imageserver-core: Removed commons-imaging and pdfbox dependencies:
foreach-custom:commons-imaging:1.0 is not on Maven Central, and turns
out to contain org/apache/commons/imaging classes. So it seems to be a
custom build of org.apache.commons:commons-imaging, that was uploaded
to the Foreach Nexus on 2013-07-27.

org.apache.commons:commons-imaging does not have a 1.0 release at the
time of writing. In fact Maven Central only has artifacts from 2019
onwards:

  https://central.sonatype.com/artifact/org.apache.commons/commons-imaging/versions

Arne's comment "Left these in for future implementations" suggests
that this is not used, and that goes for org.apache.pdfbox:pdfbox as
well (same commit). And indeed, the build passes without these
dependencies.